+++ src/main/webapp/resources/js/remit/foreignRemit.js
@@ -1,17 +1,18 @@
 "use strict"
 var foreignRemit = foreignRemit || {}
 foreignRemit = (()=>{
+	
 	const WHEN_ERR = '레미트 js파일을 찾지 못했습니다.'
-	let _,js,auth_js,main_vue_js,remit_vue_js,cookie_js,send
+	let _,js,auth_js,main_vue_js,remit_vue_js,cookie_js,amount
 
-	let init = (x)=>{
+	let init = ()=>{
 		_ = $.ctx()
 		js = $.js()
 		main_vue_js = js + '/vue/main_vue.js'
 		auth_js = js + '/cmm/auth.js'
 		cookie_js = js + '/cmm/cookie.js'
 		remit_vue_js = js + '/remit/remit_vue.js'
-		send = x
+		amount = sessionStorage.getItem('amount')
 	}
 	let onCreate =()=>{
 		init()
@@ -29,12 +30,13 @@ foreignRemit = (()=>{
 		})
 	}
 	
-	
-	
 	let setContentView = ()=>{
 		$('.themoin-main')
-		.html(remit_vue.remit_first(send))
+		.html(remit_vue.remit_first())
 		$('.themoin-footer').empty()
+		if(amount >= 3000)
+		{$('#fee_check').text('12$')}
+		else {$('#fee_check').text('6$')}
 	}
 	let remit_sec =()=>{
 		$('.themoin-main')
