+++ src/main/webapp/resources/js/remit/remit_vue.js
@@ -1,6 +1,8 @@
 var remit_vue = remit_vue || {}
 remit_vue = {
-	remit_first: x => {
+	remit_first: () => {
+			amount = sessionStorage.getItem('amount')
+
 		return `<div class="themoin-remit-extended-component"> 
 						<h2 class="title-amount">보내는 금액을 확인해주세요.</h2> 
 						<div id="moin-event-amount" class="moin-event-amount" style="text-align: center;"> 
@@ -10,7 +12,7 @@ remit_vue = {
 							<div class="amount-row"> 
 								<div class=""> 
 									<p>송금 금액</p> 
-									<input class="send-amount" type="text" tabindex="0" placeholder="0.00" value="${x}"> 
+									<input class="send-amount"  type="text" tabindex="0" placeholder="0.00" value="${amount}"> 
 								</div> 
 								<div class="unit-select disabled" tabindex="0.00"> 
 									<p>미화</p><h3>USD</h3> 
@@ -18,7 +20,7 @@ remit_vue = {
 							</div> 
 							<div class="description"> 
 								<i class="empty"></i> 
-								<p><span>12$</span> USD (수수료)</p>
+								<p><span id="fee_check"></span> USD (수수료)</p>
 								<div class="spacer"></div> 
 								<p>3,000$ 이상은 12$</p>
 							</div>
